// backend/src/modules/reminder/enums.ts
export enum ReminderMode {
  BIRTHDAY = 'BIRTHDAY',
  MEETING = 'MEETING',
  CLASS = 'CLASS',
  SCHEDULE = 'SCHEDULE',
  CUSTOM = 'CUSTOM',
}

export enum RepeatType {
  ONCE = 'ONCE',
  YEARLY = 'YEARLY',
  MONTHLY = 'MONTHLY',
  WEEKLY = 'WEEKLY',
}

export enum SendStatus {
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}